// lumi-backend/src/core/models/entities/pdf-extracted.entity.ts
export class PdfExtractedEntity {
    invoiceNumber: string;       
    clientNumber: string;
    totalValue: string; 
    referenceMonth: string;
    eletricEnergyKwh: string;
    eletricEnergyValue: string;
    sceeEnergyKwh: string;
    sceeEnergyValue: string;
    compensedEnergyKwh: string;   
    compensedEnergyValue: string;
    publicContrib: string;
}